Restructure cmd directory
The cmd directory is restructured to follow the standards as described
by https://github.com/golang-standards/project-layout/blob/master/cmd/README.md

The cmd directory just contains the main.go file for each executable
and the rest of the code is moved to pkg directory.
