--- cmd/.vscode/launch.json
@@ -1,17 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "Generate RPC",
-      "type": "go",
-      "request": "launch",
-      "mode": "auto",
-      // "program": "${fileDirname}",
-      "program": "./cli/main.go",
-      "args": ["generateRpc"]
-    }
-  ]
-}
