--- cmd/cli/main.go
@@ -1,11 +1,0 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-
-*/
-package main
-
-import "github.com/tim-pipi/cloudwego-api-gateway/cmd/cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
