+++ cmd/cloudwego/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/tim-pipi/cloudwego-api-gateway/pkg/cmd"
+)
+
+// Runs the CLI tool for the CloudWeGo API Gateway.
+func main() {
+	cmd.Execute()
+}
