+++ pkg/cmd/setIdlDir.go
